Writing program color attribute to code segment 0. Temporarily changed float registers to int, to verify color; need to write/read values via bit conversion.

+++ src/edu/smu/cse7343/bakos/os/Memory.java
@@ -9,17 +9,17 @@ package edu.smu.cse7343.bakos.os;
 
 public class Memory {
 
-    float[] registers;
+    int[] registers;
 
     public Memory(int size) {
-        registers = new float[size];
+        registers = new int[size];
     }
 
     public int totalSize() {
         return registers.length;
     }
 
-    public void write(int address, float value) {
+    public void write(int address, int value) {
         if (address >= registers.length) {
             System.out.println("TRAP: Address out of bounds."); // TODO
         } else {
